python_codes/requirements.txt
matplotlib==3.3.1
numpy==1.19.1
Pillow==7.2.0
tensorflow-gpu==2.3.0
opencv-python==4.4.0.42
scikit-learn==0.23.2
pandas==1.1.1